moves property-specific endpoints from OSS -> plus
